drivetrainmecanum serves as super class for opmodes
