+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DriveTrainMecanum.java
@@ -1,0 +1,71 @@
+package org.firstinspires.ftc.teamcode;
+
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.util.Range;
+
+import java.util.ArrayList;
+
+@TeleOp
+@Disabled
+/*
+This class serves as the OpMode class for the subsystem organization.
+ */
+public class DriveTrainMecanum extends OpMode {
+    private ElapsedTime runtime = new ElapsedTime();
+    private DcMotor leftDrive = null;
+    private DcMotor rightDrive = null;
+    private ArrayList<DcMotor> motors =new ArrayList<DcMotor>();
+
+    @Override
+    public void init() {
+        telemetry.addData("Status", "Initialized");
+
+        leftDrive  = hardwareMap.get(DcMotor.class, "left_drive");
+        rightDrive = hardwareMap.get(DcMotor.class, "right_drive");
+
+        leftDrive.setDirection(DcMotor.Direction.FORWARD);
+        rightDrive.setDirection(DcMotor.Direction.REVERSE);
+
+        telemetry.addData("Status", "Initialized");
+    }
+
+    // runs REPEATEDLY after the driver hits INIT, but before they hit PLAY
+    @Override
+    public void init_loop() {
+    }
+
+    // runs after drive hits PLAY
+    @Override
+    public void start() {
+        runtime.reset();
+    }
+
+    // runs REPEATEDLY after the driver hits PLAY but before they hit STOP
+    public void loop() {
+        double leftPower;
+        double rightPower;
+
+        double drive = -gamepad1.left_stick_y;
+        double turn  =  gamepad1.right_stick_x;
+        leftPower    = Range.clip(drive + turn, -1.0, 1.0) ;
+        rightPower   = Range.clip(drive - turn, -1.0, 1.0) ;
+
+        leftDrive.setPower(leftPower);
+        rightDrive.setPower(rightPower);
+
+        telemetry.addData("Status", "Run Time: " + runtime.toString());
+        telemetry.addData("Motors", "left (%.2f), right (%.2f)", leftPower, rightPower);
+    }
+
+    public void reset() {
+        for(DcMotor motor: motors) {
+            motor.setPower(0);
+            motor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+            motor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        }
+    }
+}
